test(maproom): SRCHREL-1005 add integration tests for quality-weighted scoring
Add 7 integration tests that validate the data setup pipeline for
quality-weighted graph scoring:
- Chunk and edge insertion workflows
- Multiple edges from same source file accumulation
- Production vs test file path creation patterns
- Config feature flags integration
- Different edge types (calls, imports)
- Moderate dataset setup performance (<2s for 50 chunks)
- Worktree scoped data

Note: Full GraphExecutor scoring tests require SQLite math functions
(ln()) not available in bundled SQLite. The algorithm logic is
validated in SRCHREL-1004 (14 unit tests) and SQL performance in
SRCHREL-0002.

+++ crates/maproom/tests/common/mod.rs
@@ -81,7 +81,7 @@ impl TestDb {
                 relpath: relpath.to_string(),
                 language: Some(detect_language(relpath)),
                 content_hash: format!("hash_{}", relpath.replace("/", "_")),
-                size_bytes: content.len() as i64,
+                size_bytes: content.len() as i32,
                 last_modified: None,
             };
             let file_id = self.store.upsert_file(&file).await?;
